refactor: clean up RscChart
